splitting levels into files
