+++ scripts/helpers/manage-saved-data.js
@@ -1,8 +1,9 @@
 import {
-  GAME_NAMES,
   STORAGE_KEY,
   TEMP_STORAGE_KEY,
 } from "../data/constants.js";
+
+import {GAME_NAMES} from "../data/game-info.js";
 
 const defaultSavedGame = GAME_NAMES.map((name, level) => ({
   level,
